initial commit
This has a very similar implementation as GitHub's scientist gem for
Ruby (github/scientist).

For the high-level design, there are a few interesting decisions to
point out. Firstly, I added a map() customizer to the existing compare()
and clean() customizers that GitHub's module exposes. You can think of
using map() with a default compare() as a pattern similar to _.sortBy(),
where you just provide a transforming function to use for each item.

A lot of functionality from GitHub's gem was stripped because I couldn't
immediately see a reason for it. We can always add it on later, but for
now, we don't need concepts like skipping experiments based on some
experiment-specific predicate or renaming the control experiment.

Perhaps the most interesting design decision is that we have promise
support built-in behind a flag. While the science module itself simply
returns the value that the control function returned (or throws the
error if the control threw), the values are then settled as promises
asynchronously to compute and emit a result. We have to keep this behind
a flag, or else you could incorrectly equate a thrown error to a
rejected one when you expected the function to be synchronous.