be more strict with observation mapping
This change forces the value going in to and coming out of the map
function to be a promise. That means values are wrapped in a promise
resolve before being given to the function, and if your function does
not return a thenable, an error is emitted.

This is useful for two reasons. For one, if only some of your behaviors
are async, you need to remember to wrap the synchronous ones in a
promise, otherwise the mapping function needs to. The second reason is
because it is easy to forget that the mapping function applies to the
promises and not their resolved values, so you might try to access an
attribute on the promise instead of using .then(), which gives you
undefined. In both cases, we're looking to reduce confusion.