add skipWhen option
This option accepts a function that returns a boolean. It is executed
before the experiment is run along with the sampler to determine if a
single instance of the experiment should be skipped in the same manner
that the sampler will skip over an experiment.

I prefer a function that returns a boolean here to a real boolean
because it lets us handle errors in a more stable way than if the user
just executed the conditional logic in the experiment setup block.