make inspect node-compatible
Because we're calling this method 'inspect', using a custom signature
that isn't the same as node's is asking for trouble. When there are test
failures and our assertion library tries to stringify the observation,
it throws an error because the 'stringify' argument is being passed an
integer for depth.

Instead of allowing for a stringification function, we'll just conform
to inspect as much as we can by using util.inspect and forwarding the
options object. That does mean that we'll reset the depth counter, but
in my opinion, it's too much complication to try and work around the
really clunky recursive inspect interface.