add ignore() configurable
This is a way to add a predicate that can conditionally ignore the
results from one or more candidates in the test. It's useful for
removing errors about cases you expect to fail due to the change. You
can either use the closure at the time the experiment is defined or you
can parameterize your condition based on the control and/or candidate
observations.

This slightly different from the GitHub implementation in that it passes
the whole observations, not the values. I think it's perfectly
acceptable to ignore errored observations too, though that does make the
interface a *little* inconsistent since all other experiment customizers
take values directly.

We are also removing ignored observations before checking if they
compare instead of only removing them if they don't match. This wasn't
an easy decision to make, but doing it this way means the comparator is
never run, so you can reason about the values coming into a comparator
more easily.