add console implementation
This is a helper that one can use via `require('scientist/console')`.
It isn't the recommended way to use the library by any means, but for
individuals curious enough to want to play with it, it provides a
no-hassle setup.

There are no tests for this because I don't know an elegant way to make
something outside of src testable. We can potentially readdress that
later, but my hope is that it won't really matter.