refactor startTime / duration into measurements
Because mapping generated a new observation, it meant the total measured
run time was extended from the first run. This is bad, as if you're
doing a few things synchronously and then mapping each, the first
observation is really measuring the time it ran for, the time all the
others ran for, and the time it took to map it. This is clearly wrong.

This refactor introduces a new `Measurement` class to help facilitate
the different kinds of timing we want to do. The default behavior
embedded in an observation is to benchmark the block. For mapping
observations, we use a helper to create a new kind of observation whose
measuring mechanism keeps the previous measurement constant. It also
means that we can use a third kind of measuring technique for settling
async promises, which _should_ be extended for the full duration, as
that is the best you can hope for with asynchronous functionality.

Start time is a strange bit of functionality that doesn't provide a lot
of use and hampers our ability to use more precise methods of timing.
We're deprecating it, and it will be removed in 2.0.